Deletion operations exceptions handled

+++ src/main/java/com/dhia/pfemanager/pfemanager/activity/task/TaskService.java
@@ -1,6 +1,7 @@
 package com.dhia.pfemanager.pfemanager.activity.task;
 
 
+import com.dhia.pfemanager.pfemanager.exceptions.EntityNotFoundException;
 import com.dhia.pfemanager.pfemanager.user.intern.Intern;
 import com.dhia.pfemanager.pfemanager.user.intern.InternRepository;
 import com.dhia.pfemanager.pfemanager.user.supervisor.Supervisor;
@@ -43,6 +44,9 @@ public class TaskService {
     }
 
     public void deleteTaskById(Integer taskId) {
+        if (!taskRepository.existsById(taskId)){
+            throw new EntityNotFoundException("This task doesn't exist");
+        }
         taskRepository.deleteById(taskId);
     }
 }
