+++ src/main/java/com/dhia/pfemanager/pfemanager/user/enterprise/EnterpriseService.java
@@ -3,6 +3,7 @@ package com.dhia.pfemanager.pfemanager.user.enterprise;
 
 import com.dhia.pfemanager.pfemanager.exceptions.EnterpriseBlockedException;
 import com.dhia.pfemanager.pfemanager.exceptions.EnterpriseEnabledException;
+import com.dhia.pfemanager.pfemanager.exceptions.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -54,6 +55,9 @@ public class EnterpriseService {
     }
 
     public void deleteEnterpriseById(Integer enterpriseId) {
+        if (!enterpriseRepository.existsById(enterpriseId)){
+            throw new EntityNotFoundException("This enterprise doesn't exist");
+        }
         enterpriseRepository.deleteById(enterpriseId);
     }
 }
